move shared url building to new file
